style: rename package to listener from event
